refactore: Optimizing and resolving form data handling
