+++ src/store/useSearchTicketStore.ts
